requirements.txt
gradio>=4.0.0
torch>=2.1.0
transformers>=4.46.0
pillow>=11.0.0
pandas>=2.2.2
numpy>=2.0.0
sentencepiece
protobuf
tqdm
